Include issues that went straight into approved state
Sometimes during API review we review related issues that were never marked
api-ready-for-review. We want to include those in the notes.

At the same time, we don't want to include all issues that ever had any API
related label applied to them because sometimes other team members apply
api-needs-work to indicate an API that isn't ready yet. Since they do so
outside of a review context we still don't want to include those in the
notes.